Refactor logic into helpers
