+++ app/views/devise/registrations/edit.html.erb
@@ -6,9 +6,7 @@
   <div class="form-inputs">
     <%= f.input :email, required: true, autofocus: true %>
 
-    <% if devise_mapping.confirmable? && resource.pending_reconfirmation? %>
-      <p>Currently waiting confirmation for: <%= resource.unconfirmed_email %></p>
-    <% end %>
+    <%= wait_email_confirmation %>
 
     <%= f.input :password,
                 hint: "leave it blank if you don't want to change it",
