+++ app/views/devise/sessions/new.html.erb
@@ -5,6 +5,7 @@
         <h2 class="title is-3">Log In</h2>
 
         <%= simple_form_for(resource, as: resource_name, url: user_session_path) do |f| %>
+          <%= f.error_notification %>
           <div class="field">
             <div class="control">
               <%= f.input :email,
