+++ config/routes.rb
@@ -1,7 +1,7 @@
 Rails.application.routes.draw do
   devise_for :users, controllers: { registrations: 'users/registrations', sessions: 'users/sessions' }
   resources :users, only: [:new, :create, :show]
-  resources :events
+  resources :events, except: [:edit, :update, :destroy]
   resources :event_attendings, only: [:new, :create]
   
   devise_scope :user do
